Use rd /s instead of shutil.rmtree on Windows in rm_rf
This hopefully fixes #861.

+++ conda/install.py
@@ -153,7 +153,10 @@ def rm_rf(path, max_retries=None):
     elif isdir(path):
         for i in range(max_retries):
             try:
-                shutil.rmtree(path)
+                if sys.platform == 'win32':
+                    subprocess.check_call(['cmd', '/c', 'rd', '/s', '/q', path])
+                else:
+                    shutil.rmtree(path)
             except OSError as e:
                 log.debug("Unable to delete %s (%s): retrying after %s seconds" %
                     (path, e, i))
